feat(vm): int is a signed 64-bit integer
An int held 32 bits, so a rowid past two billion, a unix time in
milliseconds, a file size, or 65536 * 65536 wrapped. It now holds the
signed 64-bit range from the literal through folding, the register, the
artifact, the host Value and the C boundary.

A tagged 64-bit word has no room for a 64-bit payload, so a value carries
the NaN box it always did plus a second word holding the int. Everything
else keeps its old encoding, and equality and hashing read one word for
every type but int. The .cdlb records the pair, which raises the format
version to 7. The SetInt operand stays narrow and a wider constant moves
out of its register, so an instruction is still one word. An index is
checked against the length in 64 bits before it becomes a usize, so a
position past 2^31 is out of bounds rather than a truncated read.

+++ src/engine.rs
@@ -694,7 +694,7 @@ impl Program {
 fn value_to_expr(v: &Value) -> Option<Expr> {
     Some(match v {
         Value::Null => Expr::Null,
-        Value::Int(i) => Expr::Int(*i as i32),
+        Value::Int(i) => Expr::Int(*i),
         Value::Float(f) => Expr::Float(*f),
         Value::Bool(b) => Expr::Bool(*b),
         Value::String(s) => Expr::String(SmolStr::from(s.as_str())),
